backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
passlib[argon2]
python-jose[cryptography]
pydantic
email-validator
minio
httpx
pytest
